requirements.txt
 # Google Drive API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0

# Firebase
firebase-admin==6.4.0

# Async HTTP requests
aiohttp==3.9.3
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.2

# Type hints
typing-extensions==4.9.0